requirements.txt
Flask
flask-restful
python-dotenv
mysqlclient
requests
bs4
selenium
webdriver_manager